Introduce `@RefasterSourceCompatibility` annotation and tooling
Summary of changes:
- Introduce the `@ PossibleSourceIncompatibility` annotation.
- Introduce the `RefasterReturnType` check that updates Refaster
  template methods to declare their most specific return type.
- Introduce the `RefasterSourceCompatibility` check that applies the
  `@RefasterSourceCompatibility` annotation to Refaster rules if and
  only if their template methods have return types that indicate that
  rule application may break source code.
- Introduce the `MoreASTHelpers#findDirectReturnStatements` helper
  method and simplify `JUnitValueSource` accordingly; this logic is also
  used by one of the new bug checkers.

+++ error-prone-contrib/src/main/java/tech/picnic/errorprone/refasterrules/PrimitiveRules.java
@@ -331,12 +331,12 @@ final class PrimitiveRules {
   /** Prefer JDK's {@link Long#compareUnsigned(long, long)} over third-party alternatives. */
   static final class LongCompareUnsigned {
     @BeforeTemplate
-    long before(long x, long y) {
+    int before(long x, long y) {
       return UnsignedLongs.compare(x, y);
     }
 
     @AfterTemplate
-    long after(long x, long y) {
+    int after(long x, long y) {
       return Long.compareUnsigned(x, y);
     }
   }
